Move module source files.
